+++ src/aks-preview/HISTORY.rst
@@ -12,6 +12,7 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 * Vendor new SDK and bump API version to 2023-09-02-preview.
+* Add the KataCcIsolation option to --workload-runtime.
 * Add `--network-plugin` to the `az aks update` command.
 
 0.5.168