+++ src/aks-preview/HISTORY.rst
@@ -14,6 +14,7 @@ Pending
 * Vendor new SDK and bump API version to 2023-09-02-preview.
 * Add the KataCcIsolation option to --workload-runtime.
 * Add `--network-plugin` to the `az aks update` command.
+* Update "VirtualMachines" agent pool type as Public Preview feature.
 
 0.5.168
 +++++++