+++ src/aks-preview/HISTORY.rst
@@ -12,6 +12,8 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 * Add `--network-plugin` to the `az aks update` command.
+* Add the KataCcIsolation option to --workload-runtime.
+* Update "VirtualMachines" agent pool type as Public Preview feature.
 * Add --disable-network-observability to `az aks update` cluster command.
 
 0.5.168